--- src/main/java/co/bitshifted/libs/reflex/http/HttpException.java
@@ -1,31 +1,0 @@
-/*
- *
- *  * Copyright (c) 2023-2023  Bitshift D.O.O (http://bitshifted.co)
- *  *
- *  * This Source Code Form is subject to the terms of the Mozilla Public
- *  * License, v. 2.0. If a copy of the MPL was not distributed with this
- *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
- */
-
-package co.bitshifted.libs.reflex.http;
-
-public class HttpException extends Exception {
-
-    private final int statusCode;
-    private final String errorMessage;
-
-    public HttpException(int statusCode, String errorMessage) {
-        super(errorMessage);
-        this.statusCode = statusCode;
-        this.errorMessage = errorMessage;
-    }
-
-    public int getStatusCode() {
-        return statusCode;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
