--- src/main/java/co/bitshifted/libs/reflex/serialize/JsonBodySerializer.java
@@ -1,24 +1,0 @@
-/*
- *
- *  * Copyright (c) 2023  Bitshift D.O.O (http://bitshifted.co)
- *  *
- *  * This Source Code Form is subject to the terms of the Mozilla Public
- *  * License, v. 2.0. If a copy of the MPL was not distributed with this
- *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
- */
-
-package co.bitshifted.libs.reflex.serialize;
-
-import co.bitshifted.libs.reflex.http.MimeType;
-import co.bitshifted.libs.reflex.http.MimeTypes;
-
-import java.util.Set;
-
-public interface JsonBodySerializer extends BodySerializer {
-
-    @Override
-    default Set<MimeType> supportedMimeTypes() {
-        return Set.of(MimeTypes.APPLICATION_JSON);
-    }
-}
