+++ src/main/java/co/bitshifted/reflex/exception/HttpClientException.java
@@ -1,0 +1,22 @@
+/*
+ *
+ *  * Copyright (c) 2023  Bitshift D.O.O (http://bitshifted.co)
+ *  *
+ *  * This Source Code Form is subject to the terms of the Mozilla Public
+ *  * License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ */
+
+package co.bitshifted.reflex.exception;
+
+public class HttpClientException extends Exception {
+
+    public HttpClientException(String message){
+        super(message);
+    }
+
+    public HttpClientException(Throwable th) {
+        super(th);
+    }
+}
