+++ src/main/java/co/bitshifted/reflex/serialize/JacksonJsonBodySerializer.java
@@ -1,0 +1,24 @@
+/*
+ *
+ *  * Copyright (c) 2023  Bitshift D.O.O (http://bitshifted.co)
+ *  *
+ *  * This Source Code Form is subject to the terms of the Mozilla Public
+ *  * License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ */
+
+package co.bitshifted.reflex.serialize;
+
+public class JacksonJsonBodySerializer implements JsonBodySerializer {
+
+    @Override
+    public <T> String objectToString(T object) {
+        return null;
+    }
+
+    @Override
+    public <T> T stringToObject(String input, Class<T> type) {
+        return null;
+    }
+}
