+++ src/main/java/module-info.java
@@ -1,10 +1,8 @@
 module co.bitshifted.libs.httpjfx {
-    requires javafx.base;
-    requires javafx.graphics;
     requires java.net.http;
     requires com.fasterxml.jackson.databind;
 
-    exports co.bitshifted.libs.reflex;
-    exports co.bitshifted.libs.reflex.http;
-    exports co.bitshifted.libs.reflex.serialize;
+    exports co.bitshifted.reflex;
+    exports co.bitshifted.reflex.http;
+    exports co.bitshifted.reflex.serialize;
 }
