--- src/test/java/co/bitshifted/libs/reflex/model/Foo.java
@@ -1,47 +1,0 @@
-/*
- *
- *  * Copyright (c) 2023  Bitshift D.O.O (http://bitshifted.co)
- *  *
- *  * This Source Code Form is subject to the terms of the Mozilla Public
- *  * License, v. 2.0. If a copy of the MPL was not distributed with this
- *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
- */
-
-package co.bitshifted.libs.reflex.model;
-
-import com.fasterxml.jackson.annotation.JsonIncludeProperties;
-
-import java.io.Serializable;
-
-@JsonIncludeProperties
-public class Foo implements Serializable {
-    private String name;
-    private String address;
-    private int age;
-
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public int getAge() {
-        return age;
-    }
-}
